Set minimum Go version to 1.16
This also selects a stable version for smartystreets/assertions.
