+++ .gitignore
@@ -3,3 +3,4 @@ Thumbs.db
 examples/output.json
 web/client/reports/
 /.idea
+/goconvey
