--- dependencies.go
@@ -1,4 +1,0 @@
-package main
-
-import _ "github.com/jtolds/gls"
-import _ "github.com/smartystreets/assertions"
