+++ go.mod
@@ -1,8 +1,10 @@
 module github.com/smartystreets/goconvey
 
+go 1.16
+
 require (
 	github.com/gopherjs/gopherjs v0.0.0-20181017120253-0766667cb4d1 // indirect
 	github.com/jtolds/gls v4.20.0+incompatible
-	github.com/smartystreets/assertions v0.0.0-20180927180507-b2de0cb4f26d
+	github.com/smartystreets/assertions v1.2.0
 	golang.org/x/tools v0.0.0-20190328211700-ab21143f2384
 )
