--- web/server/system/shell_older_than_1_16.go
@@ -1,7 +1,0 @@
-// +build !go1.16
-
-package system
-
-func compile(directory, gobin, tagsArg string) Command {
-	return NewCommand(directory, gobin, "test", "-i", tagsArg)
-}
